commentaryflow/requirements.txt
# CommentaryFlow dependencies
# Install: pip install -r commentaryflow/requirements.txt

# Existing Contribnote deps (already installed)
openpyxl>=3.1.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Word/PDF export
python-docx>=1.1.0
# docx2pdf>=0.1.8   # Optional — uncomment if PDF export needed (requires LibreOffice on non-Windows)
